Some placeholder work on the left widget ie. the toolbox.
I put some icons there, obviously we want them looking better (I can probably do that, I'm okay with Photoshop, or we can ask a better artist, I have a couple of contacts) but this is how I imagined the toolbar to work. Hope you like it. I was looking at the buttons in the PTE and how to sort them into different groups or tabs the best and I came up with one group for "notes" (this has the note and rest buttons for all note lengths in it and that kind of stuff), one group for "embellishments" (slide, legato and all that) and one for the actual "score" (for adding and removing sections, adding markers and so on.

Let me know what you think.

git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@17 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c