Initial commit for a CMake-based build system. This is currently only tested on Linux with Qt5 at the moment.
This should make it much easier to support other compilers (such as Clang on
OSX) and generate project files for Visual Studio. It's also much simpler to
use CMake to detect libraries like Boost, which should make it easier for
people to compile the project from source once we release publicly.


git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@911 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c