Initial work on irregular groupings.
The API will be fairly different - we no longer store information about which
positions are the start/middle/end of the group. Instead, my plan is to have an
external class that stores which positions are part of the group, and the
position itself only knows its timing (e.g. 3/2). This will make rendering and
editing much simpler, and also allow us to have nested irregular groups (which
show up in songs like Stairway to Heaven)


git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@932 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c