Fix an edge case when merging scores.
If the number of strings in the staves changes from one system to the
next in the source score, it is much simpler to just create a new system
in the destination score instead of attempting to merge those two sets
of staves in a single system. Previously it was possible for a guitar to
be assigned to a staff with more strings than its tuning, which led to
errors.

Fixes: #206