Add files via upload
This is the first upload of midifile.cpp with the else of generateGradualBends having the bend's duration * duration in there so that the midi file will send ok irregardless of whether the note is the same as the bend in duration or longer.  The bend will simply hold and ring out the duration.  So it has only one line changed from my previous versions.