Add files via upload
I condensed the if/else if/else in generateGradualBend it an if/else.  I will keep my code as it was earlier today.  That is, force a note to be the same duration as a bend.  A user can still undo the duration (for whatever strange reason), and/or increase the duration (again, for whatever strange reason).  But the note duration SHOULD match the bend's duration.  There's no other simple way to implement it any differently at this time.  The duration parameter passed into that function IS the note's duration.  Thus, I've changed the duration from bend.getDuration() * duration back to just duration as in my original post of this file.  I will have to look into what it would take to implement this to stop and hold the bend for a duration longer than the bend's duration (if this is needed).  For now, at least a bend will sound for the duration of the note, which seems like the expected behavior.