This is per Issues #224/225 - the BendAndHold bug fix of not holding.  Three files: powertabeditor.cpp/.h, and midifile.cpp
This is an update to the BendAndHold "Task" reference in Issues #224/225.  I've removed the function updateNoteDurationForBend in powertabeditor.cpp (and it's declaration in powertabeditor.h), edited the editBend method to fill in tied notes for the hold part of the bend.  And I've added a new pitchwheel method for a later stage of development (static method undeclared).  The new behavior of BendAndHold does not worry about the note the bend is added to (in keeping with the original spirit of the BendAndHold case in the switch).  But it DOES add a hold portion.