Update to Stage 2 implemenatin of Bend and Hold and Normal Bend
Updating Stage 2 with fix of Normal Bend.  Normal Bend and Bend and Hold will both not use updateNoteDurationForBend.  Had briefly put it back in, but the code sounds the same between using ties and elongating the note.  So the code for normal bend and bend and hold are now more similar and each does it's thing ok.  A future issue might be to smooth out the bends more.  Low-priority item that can be done.  So I added a switch case and separated out a case to make it easier to deal with normal bends.  It's also possible to create a normal bend by simply using 0 for additional notes in a bend and hold, as before.  This may give slightly different behavior and this might be desirable.