Change how the bundle is generated on OSX.
Use the MACOSX_BUNDLE target property to directly generate a .app bundle. This ensures that the development environment is the same as the final release

The tuning dictionary is now installed in the Resources folder, which  QStandardPaths supports

The Info.plist is now generated and uses the PTE_VERSION variable to avoid a hardcoded version number.