Refactor the ScoreLocation class to more cleanly handle const scores
These are now a separate type (similar to 'const_iterator' vs 'iterator'), which catches some errors that previously would be runtime errors.

Also refactored some awkward code in powertabeditor.cpp when handling click events, which previously was avoiding use of the non-writeable score that would raise exceptions.
Now, it just moves the caret location, which is also a precursor to being able to click / double-click on more items such as tempo markers.