Allow clickable items to be selected or double-clicked.
This is a bit of work towards improving the editing behaviour for symbols like tempo markers (e.g. currently, it's not possible to edit an existing marker - it gets removed and then you need to create a new one from scratch)

Clickable items can now be selected from single clicks, which in the future could be used to allow deletion of an item.
Double clicking will now trigger an event such as opening a dialog.

Currently this only applies to the standard clef and the tab clef, but will be expanded elsewhere in future commits.

Bug: #192