Fix issues with scaling the score info when printing.
We now rescale based on the ratio of the device width and the normal staff width, rather than rescaling each item to fit the device width, and maintain the correct x position for items that are centered.
This is needed for the score info, which might be a smaller width than the staves, and is aligned in the center.

Fixes: #338