Initial work on importing rhythm slashes from v1.7 files
The idea will be to store rhythm slashes as just a regular staff (e.g. containing notes for all of the chord notes), which hopefully will make many operations like MIDI playback straightforward. However, they'll of course be treated differently for editing and rendering to appear as rhythm slashes

With this change, any rhythm slashes in the v1.7 file result in an extra staff being added to the system, and most of the note properties from rhythm slashes are imported

Some extra bookkeeping is required to update the player changes correctly when rhythm slashes appear / disappear from one system to the next.

#16