Fix crashes when playing back notes while editing
This was caused by accessing the score from the midi thread while it was being changed by the main thread. The new approach is to generate the midi data from the main thread and hand this off to the midi thread. This is only done for the "single note" playback mode, since the normal playback has additional restriction to prevent edits during playback, but this would be a good further improvement in the future.

- Ensure that the MidiFile class is movable but not copyable

- Add a constant for the default pan value

#47