Fix a bug where shortcut preferences were not loaded successfully on macOS
This was because the settings keys for shortcuts contain the '.' character, which was also being used to encode the key's hierarchy instead of '/', so the setting was being loaded into a different path in the hierarchy. I'm not sure why this was originally done since '/' seems to still be a valid character for keys in plist files.