Stop selecting barlines from also selecting all preceding notes (#416)
This is only a partial fix for #416 as barlines are still not copied,
but that seems like a separate issue :)

The bug seems to be coming from, when creating the barlines,
it was coping the staff location & changing the position index,
but not also changing the selection start index, which is still
at 0 (hence why all preceding notes are selected).

This bug manifested in a few other places as well, so also
fixed them up.